Port part of trunk r40659 to prevent segfaults from blocksize calculation overflow.
git-svn-id: https://svn.r-project.org/R/branches/R-2-4-branch@40827 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/memory.c
@@ -1617,7 +1617,7 @@ void vmaxset(char *ovmax)
 char *R_alloc(long nelem, int eltsize)
 {
     R_size_t size = nelem * eltsize;
-    double dsize = nelem * eltsize;
+    double dsize = (double)nelem * eltsize;
     if (dsize > 0) { /* precaution against integer overflow */
 	SEXP s;
 #if SIZEOF_LONG > 4
